Alter replication protocol to use leader epoch
Currently, when a replica starts, it truncates its log up to the high
watermark (HW) to remove any potentially uncommitted messages. There are
a couple edge cases with this method of truncating the log that could
result in data loss or replica divergence.

This solves the issue by using a leader epoch rather than the high
watermark for truncation.

Fixes #38