Fix idle follower ack delay
Fix issue #102 by having the leader notify followers when there is new
data available after the follower has caught up with the leader's log
end offset.

This also introduces a new config `replica.max.idle.fetch.wait` which
controls the max amount of time an idle follower will wait before making
a replication request. The primary purpose of this is to ensure timely
check-ins with the leader for health checking.