Add kind-based dev cluster
This now uses kind to create a real dev setup with a NATS cluster and a
3-node Liftbridge cluster.

You'll need to download kind and skaffold to make this work.

Then do:
1. `make kind-up` to bring up the kind cluster
2. `make kind-apply` and then let it finish

It'll also expose both the Liftbridge nodes as well as the NATS board
on local ports so you can run the examples against that cluster.