Fix RPC propagation race edge case
Fix a race condition edge case on RPC propagation where an RPC is made
when there is no metadata leader. If the server receiving the RPC
becomes leader, it may propagate the request to itself before the
propagation NATS subscription has been established. In this case, the
propagated RPC is missed and the original RPC times out.

To fix this, we check if the server receiving the RPC has become the
metadata leader. If it has, there is no need to propagate the request as
it can be performed locally.