Send leader epoch on replication requests
Send the leader epoch on replication requests. When a partition leader
receives a replication request, check that the leader epoch is the
current leader's epoch. If not, drop the request.

This fixes an edge case where a leader is partitioned from the rest of
the ISR and a new partition leader is elected. Previously, both the old
leader and new leader would receive the replication requests. Followers
would reject responses from the old leader, but the old leader would
still commit messages due to the replication requests sent by followers
and stay in a zombie state.