Allow overriding viper config with env vars
For example with a config.yml of:

```yml
host: localhost
```

One can override the advertised host with:

```sh
env LIFTBRIDGE_HOST=liftbridge-0.liftbridge-headless \
  liftbridge --raft-bootstrap-seed --config config.yml
```

I stumbled upon this trying to run the standalone-dev container inside
docker-compose where I was using the virtual network hostnames to wire
everything together. Which lead me to these issues:

https://github.com/liftbridge-io/go-liftbridge/issues/33
https://github.com/liftbridge-io/liftbridge/issues/138

The initial connection was fine but the advertised host in the metadata
returned from liftbridge was `127.0.0.1` so when the go-liftbridge
client further went to connect to the "brokers" it would fail.

I'm pretty sure people would struggle to get this working with
statefulsets unless different config files were being injected per
replica that defined the host properly.

With this change the `LIFTBRIDGE_HOST` env var could be passed into the
pod more easily.