Fix data waiter race in replication
Fix a race condition around registering data waiters during replication
when a replica has caught up with the leader. This involved sending the
notification to the replica before unregistering the waiter, which could
cause a subsequent request from the replica from being unable to
register a new waiter since the previous one had not yet been removed.
This race would result in extremely high tail latencies on message
commits.

To fix this, we simply unregister the waiter before sending the
notification to the replica.