Fix bug related to stream deletion
This fixes a bug where when a stream is deleted, recreated, messages are
published to the recreated stream, and then the server is restarted, the
published messages are lost. This was due to the way operations were
being applied to the Raft FSM. When recovering the Raft log, we now use
a tombstone marker to mark streams for deletion and only delete them if
they have not been recreated.